backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
click==8.3.1
fastapi==0.129.0
h11==0.16.0
httptools==0.7.1
idna==3.11
loguru==0.7.3
packaging==25.0
pip==26.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.1
setuptools==80.10.2
SQLAlchemy==2.0.46
starlette==0.52.1
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
wheel==0.46.3
bcrypt==4.3.0
alembic==1.15.1
PyJWT==2.10.1
